started rudimentary implementation of generic poolboy setup
